mvc.Listener: improve docs, add callbackArguments to TS, move mvc classes into multiple files
