+++ docs/src/joint/api/mvc/Listener/prototype/callbackArguments.html
@@ -1,0 +1,7 @@
+<pre class="docs-method-signature"><code>listener.callbackArguments</pre>
+
+An array of constructor arguments.
+
+<pre><code>const listener = new mvc.Listener(callbackArg1, callbackArg2);
+listener.callbackArguments; // [callbackArg1, callbackArg2]</code></pre>
+
