+++ docs/src/joint/api/mvc/Listener/prototype/stopListening.html
@@ -1,0 +1,3 @@
+<pre class="docs-method-signature"><code>listener.stopListening()</pre>
+
+Tell the listener to remove all of its registered callbacks.
